move abi-types to crypto package
These types are not specific to nitro channel states. We expect to
re-use them for defining vouchers.
